Initial work to parse env vars stored in pass store
This provides a command line app to parse env vars who's values start
`pass:` and looks up the values of the env vars in pas store based on
the remainder of the end var, using it as the pass store path to look up
the env var value.

This then echos out all the env vars that have been processes, so that
the output can be `eval`ed.